Suggest 사용자 필요한 총 열량 구현 료

+++ src/main/java/com/witheat/WithEatServer/Domain/Dto/request/MemberHeightRequestDto.java
@@ -10,6 +10,6 @@ import java.time.LocalDate;
 @NoArgsConstructor
 @AllArgsConstructor
 public class MemberHeightRequestDto {
-    private int height;
+    private Long height;
     private LocalDate height_date;
 }
